Make generator modules an independent Soong plugin
This commit extracts `genBackend`-related code from the main Soong plugin into
an independent plugin, making the main plugin redundant now that all of its
parts have been made into independent plugins.

There is no support for generated libraries and binaries as that would introduce
maintenance issues.

This commit also includes some refactoring to the extracted code:

* `inoutForSrc` is taken from `core/generated.go` and refactored to use
  `soongInout` instead of `inout` and `android.Path` instead of
  `sourceFilePath`.
* Tags are taken from `core/generated.go` but now disambiguated using
  Go's typesystem the way it is done in Soong.
* All references to variables from `core/soong_config.go.in` are removed.

`transform_source` test is updated to use different `export_gen_include_dirs`
on Soong, since here the sources matched by regex are prefixed not relative to
bob/tests but relative to the module directory. Since these paths will be
handled correctly by the Android.bp backend there is no reason to fix this.
